64-bit tags in TALM interpreter, re-run all benchmarks, regenerate paper figures
- Changed tag fields from unsigned int (32-bit) to uint64_t in interp.h
  (oper, marker_t, opmatch_t structs)
- Updated interp.c: all tag-related functions, TAGVAL/VALTAG operations,
  exec-only sentinel checks (.value.i < 0 -> .value.li < 0), printf formats
- Fixes tag overflow for recursive depth > 9 levels (base-9 encoding)
- Re-ran all 5 benchmarks with 64-bit interpreter (20,696 runs, 0 errors):
  mergesort (2400), dyck (15120), fibonacci (144), matmul (2400), gc (720)
- Regenerated all paper figures from new data
- Added PDF output to gen_paper_fig2_fig4.py

+++ TALM/interp/include/interp.h
@@ -22,6 +22,8 @@
 #include <string.h>
 #include <errno.h>
 #include <unistd.h>
+#include <stdint.h>
+#include <inttypes.h>
 
 
 
@@ -143,7 +145,7 @@ struct oper {
 	//instr_t *dst;
 	//void *value;
        	generic_vartype value;
-	unsigned int tag; //this tag corresponds to the loop iteration
+	uint64_t tag; //this tag corresponds to the loop iteration
 	int exec; //the exec tag has two parts, the callgroup number(high bits) and the dynamic part corresponding to the instance of the function call. The dynamic part is set to the counter of the callsnd instruction that sends the operand. Everytime a callsnd instruction executes it increments its counter after sending the operand.
 	int spec; //speculation number.
 	//char isspeculative; //indicates if the operand is speculative. used for gc. 
@@ -166,10 +168,10 @@ typedef struct optoken {
 } optoken_t;
 typedef struct marker {
 	msgtype_t type;
-	unsigned int tag;
+	uint64_t tag;
 } marker_t;
 struct opmatch {
-	unsigned int tag;
+	uint64_t tag;
 	int exec;
 	int spec;
 	oper_t *op[MAX_SOURCE];
